+++ tinybrain/src/env.rs
@@ -3,7 +3,7 @@ use std::env;
 const ENV_AUTHORIZATION: &str = "AUTHORIZATION";
 const ENV_ROBOCALL_URL: &str = "ROBOCALL_URL";
 
-const DEFAULT_ROBOCALL_URL: &str = "ws://localhost:3012";
+const DEFAULT_ROBOCALL_URL: &str = "ws://127.0.0.1:3012";
 
 lazy_static! {
     pub static ref AUTHORIZATION: Option<String> = env::var(ENV_AUTHORIZATION).ok();
